show loction type in location tables, refactoring, set province values for business, refactor
