+++ packages/bcer-data-portal/app/src/constants/localEnums.tsx
@@ -147,19 +147,3 @@ export enum ReportStatus {
   Missing = 'missing',
   NotRequired = 'notRequired',
 }
-
-export enum ProvinceLabels {
-  'AB' = 'Alberta',
-  'BC' = 'British Columbia',
-  'MB' = 'Manitoba',
-  'NB' = 'New Brunswick',
-  'NL' = 'Newfoundland and Labrador',
-  'NS' = 'Nova Scotia',
-  'NT' = 'Northwest Territories',
-  'NU' = 'Nunavut',
-  'ON' = 'Ontario',
-  'PE' = 'Prince Edward Island',
-  'QC' = 'Quebec',
-  'SK' = 'Saskatchewan',
-  'YT' = 'Yukon',
-}
