+++ packages/bcer-retailer-app/app/src/constants/arrays.tsx
@@ -1,21 +1,21 @@
 export const locationTypeOptions = [
-    { value: 'physical', label: 'Physical' },
-    { value: 'online', label: 'Online' },
-    { value: 'both', label: 'Physical and Online' }
-  ];
+  { value: 'physical', label: 'Physical' },
+  { value: 'online', label: 'Online' },
+  { value: 'both', label: 'Physical and Online' }
+];
   
   export const provinceOptions = [
     {value: 'AB', label: 'Alberta' },
-      {value: 'BC', label: 'British Columbia' },
-      {value: 'MB', label: 'Manitoba' },
-      {value: 'NB', label: 'New Brunswick' },
-      {value: 'NL', label: 'Newfoundland and Labrador' },
-      {value: 'NS', label: 'Nova Scotia' },
-      {value: 'NT', label: 'Northwest Territories' },
-      {value: 'NU', label: 'Nunavut' },
-      {value: 'ON', label: 'Ontario' },
-      {value: 'PE', label: 'Prince Edward Island' },
-      {value: 'QC', label: 'Quebec' },
-      {value: 'SK', label: 'Saskatchewan' },
-      {value: 'YT', label: 'Yukon' },
-  ];  
+    {value: 'BC', label: 'British Columbia' },
+    {value: 'MB', label: 'Manitoba' },
+    {value: 'NB', label: 'New Brunswick' },
+    {value: 'NL', label: 'Newfoundland and Labrador' },
+    {value: 'NS', label: 'Nova Scotia' },
+    {value: 'NT', label: 'Northwest Territories' },
+    {value: 'NU', label: 'Nunavut' },
+    {value: 'ON', label: 'Ontario' },
+    {value: 'PE', label: 'Prince Edward Island' },
+    {value: 'QC', label: 'Quebec' },
+    {value: 'SK', label: 'Saskatchewan' },
+    {value: 'YT', label: 'Yukon' },
+];
