+++ packages/bcer-retailer-app/app/src/constants/localEnums.tsx
@@ -175,31 +175,3 @@ export enum ReportStatus {
   Missing = 'missing',
   NotRequired = 'notRequired',
 }
-
-export enum LocationType {
-  physical = 'physical',
-  online = 'online',
-  both = 'both',
-}
-
-export enum LocationTypeLabels {
-  'physical' = 'Physical',
-  'online' = 'Online',
-  'both' = 'Physical and Online',
-}
-
-export enum ProvinceLabels {
-  'AB' = 'Alberta',
-  'BC' = 'British Columbia',
-  'MB' = 'Manitoba',
-  'NB' = 'New Brunswick',
-  'NL' = 'Newfoundland and Labrador',
-  'NS' = 'Nova Scotia',
-  'NT' = 'Northwest Territories',
-  'NU' = 'Nunavut',
-  'ON' = 'Ontario',
-  'PE' = 'Prince Edward Island',
-  'QC' = 'Quebec',
-  'SK' = 'Saskatchewan',
-  'YT' = 'Yukon',
-};
